feat: MethodClient 호출하는 로직 분리
