+++ backend/src/main/java/com/pickpick/member/application/MemberInitializer.java
@@ -1,12 +1,8 @@
 package com.pickpick.member.application;
 
-import com.pickpick.exception.SlackApiCallException;
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
-import com.slack.api.methods.MethodsClient;
-import com.slack.api.methods.SlackApiException;
-import com.slack.api.model.User;
-import java.io.IOException;
+import com.pickpick.support.SlackClient;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
@@ -17,10 +13,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class MemberInitializer {
 
-    private final MethodsClient slackClient;
+    private final SlackClient slackClient;
     private final MemberRepository memberRepository;
 
-    public MemberInitializer(final MethodsClient slackClient, final MemberRepository memberRepository) {
+    public MemberInitializer(final SlackClient slackClient, final MemberRepository memberRepository) {
         this.slackClient = slackClient;
         this.memberRepository = memberRepository;
     }
@@ -28,7 +24,7 @@ public class MemberInitializer {
     @PostConstruct
     void setupMember() {
         List<String> savedSlackIds = findSavedSlackIds();
-        List<Member> currentWorkspaceMembers = fetchWorkspaceMembers();
+        List<Member> currentWorkspaceMembers = slackClient.findAllWorkspaceMembers();
         List<Member> membersToSave = filterMembersToSave(savedSlackIds, currentWorkspaceMembers);
 
         memberRepository.saveAll(membersToSave);
@@ -41,29 +37,10 @@ public class MemberInitializer {
                 .collect(Collectors.toList());
     }
 
-    private List<Member> fetchWorkspaceMembers() {
-        try {
-            return toMembers(slackClient.usersList(request -> request)
-                    .getMembers());
-        } catch (IOException | SlackApiException e) {
-            throw new SlackApiCallException("usersList");
-        }
-    }
-
     private List<Member> filterMembersToSave(final List<String> savedSlackIds,
                                              final List<Member> currentWorkspaceMembers) {
         return currentWorkspaceMembers.stream()
                 .filter(it -> !savedSlackIds.contains(it.getSlackId()))
                 .collect(Collectors.toList());
-    }
-
-    private List<Member> toMembers(final List<User> users) {
-        return users.stream()
-                .map(this::toMember)
-                .collect(Collectors.toList());
-    }
-
-    private Member toMember(final User user) {
-        return new Member(user.getId(), user.getProfile().getDisplayName(), user.getProfile().getImage512());
     }
 }
