+++ backend/src/main/java/com/pickpick/message/application/ReminderSender.java
@@ -3,10 +3,8 @@ package com.pickpick.message.application;
 import com.pickpick.exception.message.SlackSendMessageFailureException;
 import com.pickpick.message.domain.Reminder;
 import com.pickpick.message.domain.ReminderRepository;
-import com.slack.api.methods.MethodsClient;
+import com.pickpick.support.SlackClient;
 import com.slack.api.methods.SlackApiException;
-import com.slack.api.methods.request.chat.ChatPostMessageRequest;
-import com.slack.api.methods.response.chat.ChatPostMessageResponse;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -20,16 +18,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class ReminderSender {
 
-    private static final String REMINDER_TEXT_FORMAT =
-            "============================ \n " +
-                    "*리마인드 메시지가 도착했습니다*!\uD83D\uDC39 \n" +
-                    "============================ \n %s";
-    private static final String ERROR_TEXT = "";
-
     private final ReminderRepository reminders;
-    private final MethodsClient slackClient;
+    private final SlackClient slackClient;
 
-    public ReminderSender(final ReminderRepository reminders, final MethodsClient slackClient) {
+    public ReminderSender(final ReminderRepository reminders, final SlackClient slackClient) {
         this.reminders = reminders;
         this.slackClient = slackClient;
     }
@@ -43,26 +35,12 @@ public class ReminderSender {
 
         for (Reminder reminder : foundReminders) {
             try {
-                sendMessage(reminder);
+                slackClient.sendMessage(reminder);
             } catch (IOException | SlackApiException | SlackSendMessageFailureException e) {
-                log.error(ERROR_TEXT, e);
+                log.error(e.getMessage(), e);
             } finally {
                 reminders.deleteById(reminder.getId());
             }
-        }
-    }
-
-    private void sendMessage(final Reminder reminder)
-            throws IOException, SlackApiException, SlackSendMessageFailureException {
-
-        ChatPostMessageRequest request = ChatPostMessageRequest.builder()
-                .channel(reminder.getMember().getSlackId())
-                .text(String.format(REMINDER_TEXT_FORMAT, reminder.getMessage().getText()))
-                .build();
-
-        ChatPostMessageResponse response = slackClient.chatPostMessage(request);
-        if (!response.isOk()) {
-            throw new SlackSendMessageFailureException(response.getError());
         }
     }
 }
